campverse_backend/requirements.txt
django>=4.2,<6.0
djangorestframework>=3.14
django-cors-headers>=4.0
pillow>=10.0
